Popups, Loading, Spinner and Gestures

+++ Ionic/conFusion/www/js/controllers.js
@@ -212,21 +212,29 @@ menuFactory.getDishes().update({id:$scope.dish.id},$scope.dish);
     $scope.promotion = menuFactory.getPromotion().get({id:0});
 
 }])
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
-
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
     $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
+
+    $ionicLoading.show({
+      template: '<ion-spinner></ion-spinner> Loading...'
+    });
 
     $scope.favorites = favoriteFactory.getFavorites();
 
     $scope.dishes = menuFactory.getDishes().query(
-        function (response) {
-            $scope.dishes = response;
-        },
-        function (response) {
-            $scope.message = "Error: " + response.status + " " + response.statusText;
-        });
-    console.log($scope.dishes, $scope.favorites);
+      function (response) {
+          $scope.dishes = response;
+          $timeout(function () {
+              $ionicLoading.hide();
+          }, 1000);
+      },
+      function (response) {
+          $scope.message = "Error: " + response.status + " " + response.statusText;
+          $timeout(function () {
+              $ionicLoading.hide();
+          }, 1000);
+    });
 
     $scope.toggleDelete = function () {
         $scope.shouldShowDelete = !$scope.shouldShowDelete;
@@ -235,10 +243,24 @@ menuFactory.getDishes().update({id:$scope.dish.id},$scope.dish);
 
     $scope.deleteFavorite = function (index) {
 
-        favoriteFactory.deleteFromFavorites(index);
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Confirm Delete',
+            template: 'Are you sure you want to delete this item?'
+        });
+
+        confirmPopup.then(function (res) {
+            if (res) {
+                console.log('Ok to delete');
+                favoriteFactory.deleteFromFavorites(index);
+            } else {
+                console.log('Canceled delete');
+            }
+        });
+
         $scope.shouldShowDelete = false;
 
-}}])
+    };
+}])
 
 .controller('AboutController', ['$scope', 'corporateFactory','baseURL', function($scope, corporateFactory, baseURL) {
